+++ thrust/system/cuda/detail/partition.h
@@ -647,7 +647,7 @@ namespace __partition {
     typename get_plan<partition_agent>::type partition_plan = partition_agent::get_plan(stream);
 
     int tile_size = partition_plan.items_per_tile;
-    int num_tiles = (num_items + tile_size - 1) / tile_size;
+    size_t num_tiles = (num_items + tile_size - 1) / tile_size;
 
     size_t vshmem_storage = core::vshmem_size(partition_plan.shared_memory_size,
                                               num_tiles);
@@ -657,7 +657,7 @@ namespace __partition {
       return status;
 
     size_t allocation_sizes[2] = {0, vshmem_storage};
-    status = ScanTileState::AllocationSize(num_tiles, allocation_sizes[0]);
+    status = ScanTileState::AllocationSize(static_cast<int>(num_tiles), allocation_sizes[0]);
     CUDA_CUB_RET_IF_FAIL(status);
     
 
@@ -674,7 +674,7 @@ namespace __partition {
     }
 
     ScanTileState tile_status;
-    status = tile_status.Init(num_tiles, allocations[0], allocation_sizes[0]);
+    status = tile_status.Init(static_cast<int>(num_tiles), allocations[0], allocation_sizes[0]);
     CUDA_CUB_RET_IF_FAIL(status);
 
     init_agent ia(init_plan, num_tiles, stream, "partition::init_agent", debug_sync);
